- new FileExporter implemented
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@949 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/Packet.cpp
@@ -117,10 +117,10 @@ int Packet::getSize() {
   return packet->size;
 }
 
-void Packet::setPts(long long int ts){
+void Packet::setPts(int64_t ts){
   packet->pts=ts;
 }
-void Packet::setDts(long long int ts){
+void Packet::setDts(int64_t ts){
   packet->dts=ts;
 }
 long long int Packet::getPts() {
